Moved arduinosubfunctions to fonction
