+++ include/fonction.h
@@ -3,59 +3,16 @@
 
 #include "structs.hpp"
 #include "navigation.h"
-#include "arduinoSubFonction.h"
 #include "tableState.hpp"
 #include "lidarAnalize.h"
 #include "lidar.h"
 #include "constante.h"
 #include <math.h>
 
-
-//TODO : Remove
-/*
-typedef enum {
-    TAKEPLANT_INIT,
-    TAKEPLANT_FORWARD,
-    TAKEPLANT_REFORWARD,
-    TAKEPLANT_BACKWARD,
-    TAKEPLANT_TAKE,
-    TAKEPLANT_END
-} fsmtakePlant_t;
-
-typedef enum {
-    VITESSEE_INIT,
-    SOLARPANEL_INIT,
-    SOLARPANEL_FIN_INIT,
-    SOLARPANEL_SETHOME,
-    SOLARPANEL_FORWARD,
-    SOLARPANEL_PUSHFOR,
-    SOLARPANEL_BACKWARD,
-    SOLARPANEL_PUSHBACK,
-    SOLARPANEL_END
-} fsmSolarPanel_t;
-
-typedef enum {
-    PUTPLANT_INIT,
-    PUTPLANT_GOBORDER,
-    PUTPLANT_PLACE,
-} fsmTPutPlant_t;
-
-typedef enum {
-    VOLPLANT_INIT,
-    VOLPLANT_GOBORDER,
-    VOLPLANT_PLACE,
-    VOLPLANT_BACKWARD,
-    VOLPLANT_ZONEFIN,
-    VOLPLANT_FIN,
-} fsmVolPlant_t;
-
-*/
-
 typedef enum {
     INTIY_INIT,
     INTIY_BACKWARD,
 } fsminitY_t;
-
 
 typedef enum {
     SETPOS_INIT,
@@ -72,19 +29,10 @@ int returnToHome(TableState* itable,CmdAsserv* iAsser);
 void resetActionneur(CmdAsserv* iAsser, Arduino* arduino);
 void opponentInAction(TableState* itable, position_t* position);
 
-//TODO : Remove
-/*
-int turnSolarPannel(TableState* itable,CmdAsserv* iAsser,Arduino* arduino);
-int takePlant2(CmdAsserv* iAsser,Arduino* arduino,TableState*itable,int xStart,int yStart, int xEnd, int yEnd, int num_zone);
-int jardinierePutPlant(TableState* itable, CmdAsserv* iAsser,Arduino* arduino,int x,int y,int theta);
-bool allJardiniereFull(TableState* itable);
-bool DeuxJardiniereFull(TableState* itable);
-bool allStockPlanteUsed(TableState* itable);
-int VolPlante(TableState* itable, CmdAsserv* iAsser,Arduino* arduino,int x,int y,int theta);
-*/
 //TODO : New functions
 int takeStock(CmdAsserv* iAsser,Arduino* arduino,TableState*itable,int xStart,int yStart, int xEnd, int yEnd, int num_zone);
 int construct(TableState* itable, CmdAsserv* iAsser,Arduino* arduino,int x,int y,int theta);
 
+void blinkLed(Arduino* arduino,int LedNb,int periode);
 
 #endif // MYFUNCTION_H
